feat: добавлен скриншот со статистикой
