+++ homework16/src/main/java/org/example/Main.java
@@ -16,15 +16,9 @@ public class Main {
     }
 
     private static void Task1() {
-
-        MDC.put("userId", "12345");
-        MDC.put("transactionId", "67890");
-
         log.info("Some info message 1");
-        log.info("Some info message 2");
-        log.info("Some info message 3");
-
-        MDC.clear();
+        log.warn("Some warn message 2");
+        log.error("Some error message 3");
     }
 
     private static void Task2(String[] args) {
