+++ homework16/src/main/java/org/example/dto/RandomNumberDTO.java
@@ -1,0 +1,16 @@
+package org.example.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class RandomNumberDTO {
+    private int min;
+    private int max;
+    private int[] loseNumbers;
+}
